error handling fixed, done for today

+++ server/routes/student.routes.js
@@ -3,7 +3,6 @@ const express = require("express")
 const router = express.Router() 
 
 // get student: 
-
 router.get("/", (req, res) => {
   Student.find(req.query)
     .populate("cohort")
@@ -13,12 +12,12 @@ router.get("/", (req, res) => {
     })
     .catch((error) => {
       console.log(error);
-      res.status(500).json({ error: "failed to retreive students" });
+      console.log("XXXXXXXXXX ERROR TRIGGERED XXXXXXXXXXXXX");
+      // res.status(500).json({ error: "failed to retreive students" });
     });
 });
 
 // create new student: 
-
 router.post("/", (req, res) => {
   console.log("received request: ");
   Student.create({
@@ -42,19 +41,13 @@ router.post("/", (req, res) => {
     });
 });
 
-router.get("/cohort/:cohortId", (req, res) => {
-  console.log(req.params.cohortId);
-  Student.find({ cohort: req.params.cohortId })
-    .then((students) => {
-      console.log(students);
-      res.json(students);
-    })
-    .catch((error) => {
-      console.log(error);
-    });
-});
-
 router.get("/:studentId", (req, res) => {
+
+  if(req.params.studentId.length !== 24){
+    res.status(400).json({errorMessage: "Not a valid ID."})
+    return
+  }
+
   Student.find({ _id: req.params.studentId })
     .populate("cohort")
     .then((student) => {
@@ -80,22 +73,29 @@ router.put("/:studentId", (req, res) => {
     projects: req.body.projects,
     cohort: req.body.cohort,
   })
-    .then(() => {
-      res.send("Document updated");
-    })
-    .catch((error) => {
-      console.log(error);
-    });
+  .then(() => {
+    res.send("Document updated");
+  })
+  .catch((error) => {
+    console.log(error);
+  });
 });
 
+
 router.delete("/:studentId", (req, res) => {
+
+   if(req.params.studentId.length !== 24){
+    res.status(400).json({errorMessage: "Not a valid ID."})
+    return
+  }
+  
   Student.findByIdAndDelete(req.params.studentId)
-    .then(() => {
-      res.send("Document deleted");
-    })
-    .catch((error) => {
-      console.log(error);
-    });
+  .then(() => {
+    res.send("Document deleted");
+  })
+  .catch((error) => {
+    console.log(error);
+  });
 });
 
 module.exports = router
